CHANGES TO ARRANGEMENT CREATION MENU

+++ app/src/main/java/com/example/athome/data/HairdresserProvidedServiceData.java
@@ -13,7 +13,7 @@ public class HairdresserProvidedServiceData {
         return service;
     }
 
-    public float getCost() {
+    public int getCost() {
         return cost;
     }
 }
